Fix 32 bit build integer narrowing warning.

+++ src/protocols/protocol_header_sync.cpp
@@ -185,7 +185,7 @@ void protocol_header_sync::handle_event(const code& ec, event_handler complete)
     }
 
     // It was a timeout, so ten more seconds have passed.
-    current_second_ += expiry_interval.count();
+    current_second_ += static_cast<size_t>(expiry_interval.count());
 
     // Drop the channel if it falls below the min sync rate averaged over all.
     if (sync_rate() < minimum_rate_)
